Frame out new matching code
New approach to configuring messaging actions: load the matching
configuration out of an EDN file and compile it in to matching clause
regexes, then dispatch on the return of the match. This is slightly
thornier, code-wise, but will allow blabs to be configured by a pretty
clean plain-text EDN file, which is sassy as heck.

Note: this is all still WIP.